Optimization: When checking if a menu is a hook/plugin page, look up the URL in a hard-coded list of wp-admin files before actually hitting the disk. On an SSD-based system, this is about 30% faster than calling is_file(). I expect (but haven't checked) that the performance improvement would be even greater when using a HDD.
Caveat: If WordPress devs decide to remove or rename one of the hard-coded files, AME might misidentify a plugin menu as a WP core menu. However, this is very unlikely to happen in practice because:

a) Modern plugins do not use PHP files as menu slugs.
b) Old plugins would not have used filenames that conflict with built-in WP admin pages.

It seems like an acceptable risk.