tagging 1.2
This version fixes some bugs and notice errors
it gives up tinyMCE plugin previously used to add links or image in favor of the WordPress built-in ones.
fixes the slashes bug when defining custom captions for stars.
adds a link on the rating stars when not viewing the idea in its single template. When on an archive template, clicking on the stars will open the idea so that the user can rate it.