Added a "Warnings" section to Tools -> Broken Links. It shows problems that might be temporary or false positives.
It's all based on a bunch of heuristics. "Temporary" warnings that persist for 3 consecutive checks or 24 hours will be automatically reclassified as "broken". Suspected false positives will always stay as warnings.

You can turn off the "Warnings" section and all of the associated heuristics by unchecking a box on the settings page. In that case, everything that would be reported as a warning will be reported as broken instead. 

Also fixed a couple of minor bugs and UI inconsistencies.