version 1.0.1
Typo fix that caused the add_rtl_to_p function not to work correctly. this function
is using DOMDocument Class to change the HTML and add the dir=rtl and the text-align=right.
If DOMDocument is not existing a regex is initiated. However, during tests I left behind call to bad name class, so DOMDocument was never called. Now it is called.