Customizer Theme Switcher: add search/filter functionality for installed themes. This should help mitigate issues with minimal horizontal space, making it easier to find a particular theme if you know what you're looking for. Theme search is only shown if there are more than 5 installed themes, to reduce clutter when there are only a couple of themes (keeping in mind that there will be at least 3 bundled themes).
The actual search functionality could probably use improvement from someone who's better with JS, but works fine (and seems to scale well). Note that because this JS is so tightly integrated into the Customizer API (which is Backbone-like), it isn't possible to directly port the JS from the existing themes experience (which is directly Backbone-driven).