Fix various bugs in sync routine including duplicates and private posts.
Tumblr's API doesn't actually provide any count of all posts on the
blog, merely a count of all *public* posts on the blog, so there's no
way to trust that we've gotten all the posts until we actually page
through the whole thing. So in this commit, we do that. In order to
avoid paging through everything at every sync, we `break` out of the
loop after 100 posts (two pages of 50 posts each) if we remember that we
previously paged through this particular blog sometime earlier. This
also adds a bunch of nicer debug logging for the cron routines.