Squashed commit containing:
* retain exsiting user role, option for group override
* add documentation for option
* remove cache from authLdap_get_uid
* fix: LDAP groups ignored if user is in no group
* Adds Eyecandy to remove unneeded config-parts

This commit adds some javascript to remove the LDAP-Groupmapping-stuff
when no GroupMapping shall be used.

This simply hides the entries so they are not lost but simply not shown

Fixes issue #40 by sorting groups by capabilities

This sorts groups by capablities and therefore adds a user to the role
with the highest number of capabilities

Bumps version

Fixes a wrong div

Adds an error to log on dupe mailaddress

THis commit adds an error log-entry as soon as a user tries to log in
using an email-address that already exists in the database.

Bumps version

Fixes DIVs in view